+++ src/main/java/com/sxu/yusa/service/dto/OrderItemDTO.java
@@ -2,6 +2,7 @@ package com.sxu.yusa.service.dto;
 
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.Objects;
@@ -15,9 +16,9 @@ public class OrderItemDTO implements Serializable {
 
     private String productName;
 
-    private Long productPrice;
+    private BigDecimal productPrice;
 
-    private Long productQuantity;
+    private Integer productQuantity;
 
     private String productIcon;
 
@@ -41,19 +42,19 @@ public class OrderItemDTO implements Serializable {
         this.productName = productName;
     }
 
-    public Long getProductPrice() {
+    public BigDecimal getProductPrice() {
         return productPrice;
     }
 
-    public void setProductPrice(Long productPrice) {
+    public void setProductPrice(BigDecimal productPrice) {
         this.productPrice = productPrice;
     }
 
-    public Long getProductQuantity() {
+    public Integer getProductQuantity() {
         return productQuantity;
     }
 
-    public void setProductQuantity(Long productQuantity) {
+    public void setProductQuantity(Integer productQuantity) {
         this.productQuantity = productQuantity;
     }
